feat: rewrite & round 2 challenges
