+++ README.md
@@ -1,2 +1,19 @@
-# csd-ctf
-CSD Write Up's and Challenges from the Cyberstudents discord. You can join at discord.gg/cyberstudents. If there are any errors or you would like to put your own writeup or solve scripts make a PR and I will add it.
+# CyberStudents CTF Archive
+
+This repository contains challenges from previous rounds of the CyberStudents CTF, along with official write-ups.
+
+The CSD-CTF is a cybersecurity capture-the-flag competition focused on students, organized by the CyberStudents Discord
+community. It features a wide range of categories, from web exploitation to OSINT, and is designed to be
+beginner-friendly. Each round lasts for around a month, and top players on the leaderboard may receive prizes. If you're
+interested in participating, you can [join us on Discord](https://discord.gg/cyberstudents-916144903686336513).
+
+If you have any questions or concerns, feel free to reach out to us on Discord or open an issue. If you spot any errors
+or have suggestions for improvements, you can open an issue or pull request.
+
+The [CyberStudents CTF Discord bot](https://github.com/cyberstudentsacademy/cyberstudents-ctf) is used to manage the
+CTF, and the source code is available on GitHub.
+
+## Rounds
+
+- [Round 1](rounds/1/README.md) (01/14/2024 - 01/17/2024)
+- [Round 2](rounds/2/README.md) (02/04/2024 - 02/26/2024)
