--- Round 1 /Loggy Log's/Challenge.md
@@ -1,7 +1,0 @@
-# Loggy log's 
-
-Category: Misc/Logs
-
-Description: Agent, we found the logs in the hacker's computer can you find anything unusual?
-
-Tip: look around the log to find the flag
