--- Round 1 /Loggy Log's/Writeup.txt
@@ -1,7 +1,0 @@
-Loggy log's 
-
-- How to solve
-Inside the file you need to look for anything unusual, and we see this "ZmFrZXtmbGFnfQ:69" but when decoded from base64 it is a fake flag. 
-Eventually we find this string "Y3Nke3lvdV9mb3VuZF9tZTQ5MjR9" which decoded gives us the flag.
-
-Flag: csd{you_found_me4924}
