--- Round 1 /Nokia's Secret/Challenge.md
@@ -1,5 +1,0 @@
-# Nokia's secret
-
-Category: Forensics/Google
-
-Description: find what is hiding in the image
