--- Round 1 /Pokemon PW Crack/Challenge.md
@@ -1,7 +1,0 @@
-# Pokemon PW Crack   
-
-~~10 ~~25 points
-
-Briefing: We have found the hash of our target, but cannot decrypt it can you? We know that his password is his favorite pokemon + a random 2 digit number Ex. pokemon69. Can send us the password?
-
-Hash:  ```76312B45FAAC2AF17B66176A4C440BBA```
