--- Round 1 /Webhunt/Challenge.md
@@ -1,11 +1,0 @@
-# Webhunt
-
-100 points
-
-Category: Web/RE
-
-Briefing: Agent, they took our best man vip3r hostage and gave us a challenge. They will free him if we find the flag in their website. Good luck.
-
-Attachments: https://437095c8-c22d-4b63-bf98-9b34cbcdc2dd-00-1r8uurngzj33.picard.replit.dev/
-
-Credits to kshau for fixing some stuff
