--- Round 1 /Where's vip3r?/Challenge.md
@@ -1,5 +1,0 @@
-# Find vip3r
-Category: OSINT
-Description: @vip3r went missing, but he sent us an image. Can you find where he is? 
-
-Tip: when submitting the flag, make sure all the spaces in the name are replaced by underscores, all the words are lowercase and we are looking for the Google maps name of the location. Ex. csd{vip3r_location_is_here}
