+++ rounds/1/Find vip3r/challenge.md
@@ -1,0 +1,30 @@
+# Find vip3r
+
+- Published: 01/15/2024 (#3/5 in round)
+- Category: OSINT
+- Points: 45
+- Author: Vipin
+
+Description: @vip3r went missing, but he sent us an image. Can you find where he is?
+
+Tip: when submitting the flag, make sure all the spaces in the name are replaced by underscores, all the words are
+lowercase and we are looking for the Google maps name of the location. Ex. csd{vip3r_location_is_here}
+
+## Attachments
+
+- [findme.png.zip](findme.png.zip)
+
+## Write-up
+
+<details>
+<summary>Reveal write-up</summary>
+
+Using a tool like Google image barely gives results, when the image is put through a EXIF viewer you won't see any
+metadata, but a little searching about the area finds us a place called Maheswara Maha Pyramid kadthal. with the same "I
+love Patriji" on Google earth
+
+Flag: csd{maheswara_maha_pyramid_kadthal} or csd{maheswara_maha_kadthal_pyramid}
+
+</details>
+
+Write-up by Vipin
