+++ rounds/1/README.md
@@ -1,0 +1,15 @@
+# Round 1
+
+This round featured 5 challenges lasting from 01/14/2024 to 01/17/2024.
+
+| #   | Name                                                  | Category         | Published  | Points | Author |
+| --- | ----------------------------------------------------- | ---------------- | ---------- | ------ | ------ |
+| 1   | [Pokemon PW Crack](Pokemon%20PW%20Crack/challenge.md) | Misc             | 01/14/2024 | 25     | Vipin  |
+| 2   | [Nokia's secret](Nokia%27s%20secret/challenge.md)     | Forensics/Google | 01/14/2024 | 10     | Vipin  |
+| 3   | [Find vip3r](Find%20vip3r/challenge.md)               | OSINT            | 01/15/2024 | 45     | Vipin  |
+| 4   | [Loggy log's](Loggy%20log%27s/challenge.md)           | Misc/Logs        | 01/16/2024 | 35     | Vipin  |
+| 5   | [Webhunt](Webhunt/challenge.md)                       | Web/RE           | 01/17/2024 | 100    | Vipin  |
+
+## Final leaderboard
+
+The leaderboard could not be finalized due to internal issues. (The final leaderboard was incomplete.)
