+++ rounds/2/Find Vip3r Pt. 2/challenge.md
@@ -1,0 +1,30 @@
+# Find Vip3r Pt. 2
+
+- Published: 02/05/2024 (#2/11 in round)
+- Category: OSINT
+- Points: 50
+- Author: Vipin
+
+Vip3r went on vacation without telling us, can you find him?
+
+Tip: send the general location of where this is. Ex. csd{lake_superior}
+
+## Attachments
+
+- [findvip3r2.png.zip](findvip3r2.png.zip)
+
+## Write-up
+
+<details>
+<summary>Reveal write-up</summary>
+
+A
+[Google Lens search](https://lens.google.com/search?ep=gisbubb&hl=en&re=df&p=AbrfA8rDKih8c8FNUpz-tHMepNUNVMjgxO_WFJ_MK9k6Cdc93jOSLPRL4RWPRyxY6IPBeOoE8K3ThXBvEimr1pph2ApBwabEyzH5GJKvxS1hsm5Ca41MPLOluIZF65LdZXpO08DwuSa64IQpKQFq-b5-eb9A1lhs_EL2EA6Ux47M5TIJ6M7UBuG2XXIoLq6MKOD1BJENtqFT0iPVFMhf5opgtVSuo81qqFYMC45fhZdYNAbUmfs0KAA2VQfWBnkaEYhUMtVjKJ7qAfQEf6Qwte1VcY4%3D#lns=W251bGwsbnVsbCxudWxsLG51bGwsbnVsbCxudWxsLG51bGwsIkVrY0tKRFl5TXpZMllUWmlMVEV5TkdVdE5ESXdNeTA1TURObUxUWmtOVE13T0RGa09HTmhOeElmYXpoa2VFbHpUbXRQVG5kV1VVVjFlRFl4WHpJd1dFUTRTRk14UXpWU1p3PT0iXQ==)
+(reverse image search) of the image gives us the location of the beach, Pensacola Beach. Format the name of the location
+as per the tip.
+
+Flag: csd{pensacola_beach}
+
+</details>
+
+Write-up by GodderE2D
