+++ rounds/2/Food Break/challenge.md
@@ -1,0 +1,39 @@
+# Food Break
+
+- Published: 02/14/2024 (#9/11 in round)
+- Category: PWD Cracking
+- Points: 30
+- Author: Vipin
+
+Have a break from solving broken (temporarily) ctf challenges :D
+
+Wrap answer in csd{}
+
+## Attachments
+
+- [hash.txt.zip](hash.txt.zip)
+
+## Hint
+
+This hint costed 15 points.
+
+<details>
+<summary>Reveal hint</summary>
+
+Try Hashes.com or Hashcat/JTR with rockyou.txt
+
+</details>
+
+## Write-up
+
+<details>
+<summary>Reveal write-up</summary>
+
+A popular website to decrypt hashes is [hashes.com](https://hashes.com/en/decrypt/hash). Pasting in the hash, we can see
+that it corresponds to `friedchicken` in SHA-512.
+
+Flag: `csd{friedchicken}`
+
+</details>
+
+Write-up by GodderE2D
