+++ rounds/2/README.md
@@ -1,0 +1,21 @@
+# Round 2
+
+This round featured 11 challenges (excluding the Sanity Check) lasting from 02/04/2024 to 02/26/2024.
+
+| #   | Name                                                            | Category                   | Published  | Points | Author        |
+| --- | --------------------------------------------------------------- | -------------------------- | ---------- | ------ | ------------- |
+| 1   | [Kitty Cat Zip Crack](Kitty%20Crack%20Zip%20Crack/challenge.md) | OSINT/PWD Crack            | 02/04/2024 | 125    | Vipin         |
+| 2   | [Find Vip3r Pt. 2](Find%20Vip3r%20Pt.%202/challenge.md)         | OSINT                      | 02/05/2024 | 50     | Vipin         |
+| 3   | [Heard of Hex?](Heard%20of%20Hex%3F/challenge.md)               | Cryptography/Steganography | 02/06/2024 | 70     | Vipin         |
+| 4   | [Out of my way](Out%20of%20my%20way/challenge.md)               | Web                        | 02/07/2024 | 55     | kshau & Vipin |
+| 5   | [Vig you!](Vig%20you%21/challenge.md)                           | Cryptography               | 02/08/2024 | 80     | Vipin         |
+| 6   | [Brute For Loot](Brute%20For%20Loot/challenge.md)               | Misc                       | 02/09/2024 | 90     | Vipin         |
+| 7   | [MultiClicking](MultiClicking/challenge.md)                     | Web Exploitation           | 02/10/2024 | 100    | kshau         |
+| 8   | [Shipwreck Collision](Shipwreck%20Collision/challenge.md)       | Cryptography               | 02/12/2024 | 100    | GodderE2D     |
+| 9   | [Food Break](Food%20Break/challenge.md)                         | PWD Cracking               | 02/14/2024 | 30     | Vipin         |
+| 10  | [Fragile Memory](Fragile%20Memory/challenge.md)                 | Reverse engineering        | 02/17/2024 | 100    | GodderE2D     |
+| 11  | [filler chall 1](filler%20chall%1/challenge.md)                 | Misc                       | 02/26/2024 | 20     | Vipin         |
+
+## Final leaderboard
+
+The leaderboard could not be finalized due to technical issues. (Tiebreakers were not handled properly.)
